Outline models and their construction
This is the first step in making the importing of data from a json file
actually work. The task parses the source file and then creates all the
resources.

Obviously this will change, but I'm just getting pieces in place.

Nothing is actually imported to Solr yet. That comes later.